--- src/AnalysingData.java
@@ -1,14 +1,0 @@
-public class AnalysingData {
-
-    //This class is for messing about for reading in the data and looking at it etc.
-    private String f1;
-    private String f2;
-
-
-    public AnalysingData(String f1, String f2){
-        this.f1 = f1;
-        this.f2 = f2;
-        System.out.println("analysis");
-    }
-
-}
